Enable the @typescript-eslint/no-unused-expressions eslint rule and fix the resulting code errors
This commit also combines affected similar implementations for cell option mapping into util classes.
