+++ databrowser/src/domain/cellComponents/components/cells/tagCell/TagCell.vue
@@ -20,8 +20,8 @@ SPDX-License-Identifier: AGPL-3.0-or-later
 <script setup lang="ts">
 import { computed, ref, toRefs, useAttrs } from 'vue';
 import EditListCell from '../../utils/editList/EditListCell.vue';
-import { useMapper } from './mapper';
 import TagTable from './TagTable.vue';
+import { useAttributeMapper } from './useAttributeMapper';
 
 const props = withDefaults(
   defineProps<{
@@ -40,7 +40,7 @@ const { tags, sortByLabel, unique } = toRefs(props);
 
 const attrs = useAttrs();
 
-const { options } = useMapper(ref(attrs), sortByLabel);
+const { options } = useAttributeMapper(ref(attrs), sortByLabel);
 
 const enableUniqueValue = computed(() => {
   // if unique is a boolean, return it
